// frontend/libs/aas-designer-core/src/lib/v3-editor/model/cm-type-option.ts
export enum CMTypeOption {
  Shells = 'Shells',
  Shell = 'Shell',
  AllSubmodels = 'AllSubmodels',
  Submodel = 'Submodel',
  ConceptDescriptions = 'ConceptDescriptions',
  ConceptDescription = 'ConceptDescription',
  SupplementalFiles = 'SupplementalFiles',
  MissingSubmodel = 'MissingSubmodel',
  Assets = 'Assets',
  Asset = 'Asset',
  SupplementalFile = 'SupplementalFile',
  Property = 'Property',
  MultiLanguageProperty = 'MultiLanguageProperty',
  SubmodelElementCollection = 'SubmodelElementCollection',
  SubmodelElementList = 'SubmodelElementList',
  Operation = 'Operation',
  Range = 'Range',
  Entity = 'Entity',
  Event = 'Event',
  Blob = 'Blob',
  File = 'File',
  ReferenceElement = 'ReferenceElement',
  SubmodelElement = 'SubmodelElement',
  Capability = 'Capability',
  RelationshipElement = 'RelationshipElement',
  AssetAdministrationShells = 'AssetAdministrationShells',
  AssetAdministrationShell = 'AssetAdministrationShell',
  SnippetGroup = 'SnippetGroup',
  ElementGroup = 'ElementGroup',
  ContactInformation = 'ContactInformation',
  AddGroup = 'AddGroup',
  OptionalElements = 'OptionalElements',
  FromJson = 'FromJson',
  SubmodelTemplate = 'SubmodelTemplate',
  RepoSubmodelTemplate = 'RepoSubmodelTemplate',
  SMFromJson = 'SMFromJson',
  ElFromJson = 'ElFromJson',
  CdFromJson = 'CdFromJson',
  Addon = 'Addon',
}